refactor: separate table from CollectorServiceAccountSection
